稍微修改所有的电力参数数据计算 by PgNat

+++ src/main/java/power/api/controller/responseModel/powerMonitoring/FrequencyResponse.java
@@ -13,5 +13,5 @@ import java.sql.Timestamp;
 @Data
 public class FrequencyResponse {
     private Timestamp createAt;
-    private double frequency;
+    private float frequency;
 }
